+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - RPC errors do not always include the root cause. For example, some gateway error messages are not output when pathfinder forwards the request.
 - RPC trace object uses wrong property `reverted_reason` instead of `revert_reason`
 
+### Added
+
+- Added the ability to concurrently process RPC batches, see the `rpc.batch-concurrency-limit` CLI argument.
+
 ## [0.9.3] - 2023-10-16
 
 ### Fixed